+++ src/pages/AdminUserPage.jsx
@@ -10,8 +10,13 @@ import Col from "react-bootstrap/Col";
 //API
 import { getAllUsers } from "../apis/admin";
 
+import { useAuth } from "../contexts/AuthContext";
+import { useNavigate } from "react-router";
+
 const AdminUserPage = ({ user }) => {
   const [allUsers, setAllUsers] = useState([]);
+  const { isAuthenticated } = useAuth();
+  const navigate = useNavigate();
 
   useEffect(() => {
     const getAllUsersAsync = async () => {
@@ -25,6 +30,12 @@ const AdminUserPage = ({ user }) => {
     };
     getAllUsersAsync();
   }, []);
+
+  useEffect(() => {
+    if (!isAuthenticated) {
+      navigate("/admin");
+    }
+  }, [navigate, isAuthenticated]);
 
   return (
     <Container>
